docker/api/requirements.txt
setuptools
wheel
aniso8601==9.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
flask-restful-url-generator==0.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
monero==0.8
mysql==0.0.2
mysql-connector-python==8.0.23
mysqlclient
protobuf==3.14.0
pycryptodome==3.10.1
pymongo==3.11.3
pyserial==3.5
pysha3==1.0.2
python-bitcoinrpc==1.0
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
varint==1.0.2
Werkzeug==1.0.1
